+++ RELEASE-NOTES.md
@@ -7,6 +7,7 @@
 * Add node_uri param for miners that supports TLS
 * Add exploder api block_info_by_height
 * Update jrocksdb to 7.2.2
+* Add MinerReport utility
 
 ## 2.0